Replace Flask for FastAPI
* Add FastAPI dependencies and removes Flask

* Renames 'server.py' to 'api.py'

* Updates index.html to call localhost:8000 instead
