+++ module6-stream-processing/risingwave/pyproject.toml
@@ -11,13 +11,13 @@ authors = [
 requires-python = ">=3.10,<3.12"
 
 dependencies = [
-    "Flask>=3.0.2,<4.0",
-    "Flask-Cors>=4.0.0,<5.0",
+    "fastapi>=0.110.0",
     "polars>=0.20.15,<1.0",
     "pandas>=2.2.1,<3.0",
     "pyarrow>=15.0.1,<16.0",
     "psycopg2-binary==2.9.9",
     "confluent-kafka>=2.3.0,<3.0",
+    "uvicorn>=0.28.0",
 ]
 
 [tool.pdm.dev-dependencies]
