+++ module6-stream-processing/risingwave/webapp/index.html
@@ -61,7 +61,7 @@
   }
 
   async function query_busiest_zones() {
-    const response = await fetch('http://127.0.0.1:5000/get_busiest_zones')
+    const response = await fetch('http://127.0.0.1:8000/get_busiest_zones')
     return await response.json();
   }
